Rewrite GraphQL subscription test
Prior to this commit, the test would require to manually start the
application and test subscriptions with hand-crafted queries using only
an HTTP client.

This commit rewrites the test as a `@GraphQlTest`, which does not
require a live server and uses instead a `GraphQlTester` for useful
assertions.

This behavior can be checked against the live server still using the
graphiql UI at http://localhost:8080/graphiql